05_debian_theme: cache background picture if not in /boot/grub filesystem
All files needed to display the boot menu, including the
background picture, should be in the same filesystem as
/boot/grub.

Indeed nothing guarantees that GRUB can actually read files
outside the /boot filesystem even if it supports it, because
it relies on platform firmware which may have limitations such
as not being able to handle other media or full media capacity.
This is what a separate /boot was designed for in the first
place, and it is the responsibility of the administrator to put
/boot in a location readable by the platform firmware.

So cache the background picture in /boot/grub not only if it
is not readable by GRUB but also if it is not in the same
filesystem as /boot/grub.

This also avoids prompting for an extra passphrase when /boot
and / are on different encrypted volumes, and allows to use
a GRUB built with minimal filesystem and abstraction support.